Enter and exit AcquisitionMode on every generator yield
This fixes the incomplete frame issue but removes the performance gains from using a generator expression.
Ideally instead it should be figured out, why using a single frame as buffer is not able to produce stable synchronous grabbing of images.
Another option would be to only roll back the state machine as far as really necessary to win back some of the performance that is lost with this change.

git-svn-id: https://ahr-svn.avtnet.local/svn/host/Products/VimbaSDK/VimbaPython/branches/PYTHON-48_incomplete_frames_in_sync_grab@28764 a56fe990-022f-c24f-abd5-7dd9c0b867e8