Access the actual .raw data of the buffer instead of parsing them via .value
Using .value truncates the data in the buffer at the first 0-byte. This might be handy for c-style-strings, but is definitely not the most common use-case for this function. Other data that might be retreived with this function may containt 0-bytes somewhere in the middle and therefore truncation will lead to loss of data.

git-svn-id: https://ahr-svn.avtnet.local/svn/host/Products/VimbaSDK/VimbaPython/branches/fix_OCT-3035@29505 a56fe990-022f-c24f-abd5-7dd9c0b867e8