Fail unittest run if any subprocess call results in error
Without this calls to flake8, mypy and coverage can fail with errors without marking the python script run as an error. As the entire test suite is run via coverage to get code coverage from it, failing unit tests only resulted in a low coverage, not an actual error result in the unit test run. This change should now properly fail the test run and therefore mark the CI pipeline in Jenkins as failed.

git-svn-id: https://ahr-svn.avtnet.local/svn/host/Products/VimbaSDK/VimbaPython/branches/improve_static_tests@30214 a56fe990-022f-c24f-abd5-7dd9c0b867e8