Switch from xmlrunner to unittest-xml-reporting for Python Unittest reporting
The xmlrunner package is no longer maintained. The new package is a drop in replacement.
This was necessary because the subtests introduced with the AllocAndAnnounce implementation are not supported by the old package. The concept of subtests generally does not map well to XML test reports. This is also mentioned in the documentation of the new package. But the new package at least has some basic support for them.

git-svn-id: https://ahr-svn.avtnet.local/svn/host/Products/VimbaSDK/VimbaPython/trunk@32036 a56fe990-022f-c24f-abd5-7dd9c0b867e8