Update capture state machine
This ensures, that frames are queued before AcquisitionStart is executed on the camera. It also modifies the order of operations during startup (VmbCaptureFrameQueue is now called before VmbCaptureStart) as this leads to a more symmertrical order of calls. The order if these function calls is not specified in the GenTL standard and was smoke tested with USB and GigE cameras. Similar changes in order of operations will be made for our VimbaC examples. VimbaCPP convenience functions and the Vimba Viewer.

Related Jira tickets: OCT-3363

git-svn-id: https://ahr-svn.avtnet.local/svn/host/Products/VimbaSDK/VimbaPython/trunk@32242 a56fe990-022f-c24f-abd5-7dd9c0b867e8