Add VimbaTestCase class
By inheriting from this base class, the functionality of
`assertNoRaises`, `get_test_camera_id` and `set_test_camera_id` do not
need to be monkeypatched into the `unittest.TestCase` class anymore.
This also now allows e.g. the Visual Studio Code test explorer to detect
test cases and run them with a press of a button simplifying development
with the help of the existing tests.

Unfortunately this introduces the need for the ugly hack of adding the
parent directory to the import path for every test file to get to the
`VimbaTestCase` class.