`convert_pixel_format` now returns result
The old implementation attempted to change the pixel format of a frame
"in place". This led to issues if users wanted to reuse a frame for
further transmissions after using the `convert_pixel_format` method on
it.

With this change the `convert_pixel_format` method no longer changed the
frame instance it was called on. Instead the result of the
transformation is returned by the method. This results in a new
allocation of memory for the transformation result. After transformation
it is safe for the user to requeue the original frame for further
transmissions and the user can keep working on the transformation
result.

Fixes OCT-3156