Add handling for Reachable/Unreachable events
If a Reachable/Unreachable event is triggered, the permitted access
modes (which are cached at creation of the Camera object) are updated.

note: This commit implements the update of the permitted access modes by
simpy updating the entire `CamerInfo` member. This should be safe but
some TODO comments are added to verify this before final merge of this
branch.