Add test for `Feature` write access reporting
When streaming mode is entered the write access to some device features
is disabled. This test ensures, that the `Feature` instance correctly
reports the write access status.

warning: Currently uses the devices `Height` feature which at teh time
of writing this test becomes read-only when streaming is enabled. If in
the future devices support changing the `Height` feature during
streaming, this test needs to be updated to use a different feature.

Related Jira tickets:
- OCT-3167
- OCT-3383