Read device id from env variable if available
This allows the user to set the environment variable
`VIMBAPYTHON_DEVICE_ID` to define the camera they want to use for
running the test cases. This can make sense if e.g. multiple GigE
cameras are found by the system because they are connected to a shared
test network.

For the test explorer within VScode setting the env variable can e.g. be
done with the help of a file named `.env` in the workspace root. The
Python extension of VScode will load environment variables from that
file. Documentation regarding the `.env` file can be found at [1].

From the commandline it is possible to either set the environment
variable for the terminal, or define it temporarily for the command that
runs the test case.

Example for CLI use on windows. Execute in the `Tests` directory:
```
$ set VIMBAPYTHON_DEVICE_ID=DEV_1AB22C00041C
$ python -m unittest -v real_cam_tests.camera_test.CamCameraTest.test_camera_streaming
```
For Linux adjust the way to define the environment variable accordingly.
This will run only the specified test case.

[1] https://code.visualstudio.com/docs/python/environments#_environment-variables