Only update `permittedAccess` in `VmbCameraInfo`
VimbaC will ensure, that the `VmbCameraInfoQuery` returns an appropriate
camera info for the requested device. Checking the device id in the
returned `VmbCameraInfo` is not necessary.

The only field in the returned `VmbCameraInfo` that changes is the
`permittedAccess` field. This is the only field that needs to be
refreshed in the cached `VmbCameraInfo`.