Add use of `convert_pixel_format` to example
The example now no longer relies on the camera directly supporting an
opencv compatible pixel format, but instead tries to select it by the
following logic.

1. If compatible format is supported by camera, use it
2. If a color format the camera supports can be transformed into the
   desired opencv format, use that
3. If a mono format the camera supports can be transformed, use that

In case 2 or 3 a transformation of the recorded frame is necessary. This
was added to the `Handler.__call__` method and is a good example of why
`convert_pixel_format` should not change in place. before the frame
would have to be transformed back to the original format in which the
camera records frames. After the previous changes from this branch, a
copy of the frame is returned so the original `frame` object in the
callback can simply be queued back to the camera with no further
worries.