Update `_load_under_windows`
This now takes the search directory for VmbC.dll from the environment
variable `VIMBAX_HOME` (capitalization does not matter). The variable is
assumed to point to a directory that itself contains a directory named
`Bin`. In this bin directory, the `VmbC.dll` should be found.

WARNING: This variable currently must be set manually since no installer
exists that takes care of this!