Update expected library versions
Questions: Why is the expected version of VmbImageTransform not checked
during load time like it is for VmbC? Should this be done? Does it make
sense to have the expected VmbC version in the codebase twice (once in
`EXPECTED_VIMBA_C_VERSION` and once in the test case change dby this
commit?