Change function name in image transform library
The change to `VmbGetImageTransformVersion` was never merged to the
`dev` branch of the image transform library. This commit reverst the
change back to using the function name as it can be currently found on
the `dev` branch. If the name ever changes in the future this commit can
simply be reverted.