Update docstrings and remove wrong exceptions
The `raise`s that are removed in this commit were not really sensible.
They are raising an `InterfaceError` even though this should rather be a
(as of yet non-existing) `TransportLayerError`. Also the case where the
transport layer is not known to `VmbSystem` is very hard to produce as
the constructor will only be called by our tl discovery mechanism and if
the user tries to do it manually, they will run into issues where they
would need to pass a valid handle...