Fix circular import type checking issue
This allows importing types in a way that would usually lead to circular
imports. The `TYPE_CHECKING` variable is defined as `False`, but type
checkers such as mypy simply set them to `True` for the purpose of
evaluating the type hints. See [1] for more info.

The import from `__future__` is done, so that types can be used directly
in the annotations and do not need to be set in quotes. This behaviour
will (most likely) become the new default in Python 3.11. See PEP 563 at
[2] for more info on the full impact of this change.

References:
[1] https://docs.python.org/3/library/typing.html?highlight=typing#typing.TYPE_CHECKING
[2] https://peps.python.org/pep-0563/