Update frame callback definition
All frame callbacks now need to have three parameters. The newly
introduced parameter at position 2 is the handle to the newly introduced
`Stream` class. As the TODO comment indicates, VmbPy will need to
resolve that handle to the actual Stream instance to make it convenient
to use for the user.