Restructure discovery functions (tl/interface/cam)
This moves the discovery functions for transport layers, interfaces, and
cameras into the `VmbSystem` class as private methods. This makes it
easier to avoid circular imports.

In the process of this migration, the types that are used to keep track
of transport layers and interfaces also changed. Before, `VmbSystem`
simply kept a list of instances for each class. Now dictionaries are
used instead. The key for each dict entry is the `VmbC` handle of the
transport layer/interface and the value is the instance of the
corresponding `VmbPy` class. This still makes enumeration of the
available instances easy, but also allows easy identification by the
handle. This is important, because VmbC identifies the transport
layers/interfaces by their handle and reports these in the C-structs
available.