Fix `VmbFrame` field types
The incorrect type for `imageData` lead to an issue, where the frames
`receiveFlags` were not interpreted correctly meaning the API though
things like the FrameID where not correctly set.

The pointer will need to be updated in the deepcopy routine of frame.
For that the offset that `imageData` has from the original `buffer`
probably needs to be calculated and an appropriate pointer must be set
in the copied frame instance relative to the new buffer there.