Make use of `imageData` pointer in `as_numpy_ndarray`
The `_buffer` variable does not necesasarily start with the image data.
With the migration to VmbC the start of the buffer may contain
additional data. The `imageData` pointer in the `VmbFrame` struct points
to the real beginning of the image data. Currently these pointers are
identical for all cameras I have tested this with, but this might change
in the future.