Improve `VmbFrame.__repr__`
The output now prints the hex formatted adresses of the buffer and the
hex formatted adress that `imageData` point to. This is likely the most
interesting information that can be gotten from those variables by
printing them as it allows the user to see the offset between beginning
of the buffer and beginning of the image data inside the buffer.