Change context decorators into functions
The implementation of the decorator as a class was probably done to keep
consistency with the `RuntimeTypeCheckEnable` decorator, that needs some
class variables to function. However the simple decorators changed here
do not rely on such class variables to work. By changing them to
functions, the migration the feature access functionality will become
easier, because decorators can be added further down the inheritance
chain more concisely.