Add `FeatureContainer` class
Also set up inheritance from that class for
- `VmbSystem`
- `TransportLayer`
- `Interface`
- `Camera` (equivalent to remote device)
- `LocalDevice`
- `Stream`

This required that the member variable `self.__feat` be renamed to
`self._feat` as otherwise name mangling would take place and complicate
things. This should not have any impact as member variables marked with
underscore should still be considered private.

For `VmbSystem` this commit also already adds the
`raise_if_outside_context` decorator to the appropriate methods.
Investigation on the other classes is still outstanding where this
decorator must be added.