Rename internal handle variable to single underscore
This is in preparation to using the `FeatureContainer` parent class to
automatically detect/attach/remove feature accessors to the class
instance. With the double underscore in front of the member variable
name, Python performs name mangling which essentially prevents using the
member variable from a different class. This was the desired behaviour,
but makes the inheritance more difficult to implement decently. Having a
single underscore in front of the member variable name still indicates,
that it is a private variable that users should not rely on, but makes
it more easily accessible.