Add feature accessor functionality to FeatureContainer
The context manager will make sure, that features are automatically
detected and attached to the instance as appropriate. For the classes
that rely on a context manager themselves (`VmbSystem` and `Camera`) the
context manager of `FeatureContainer` is used at those context manager
enter/exit methods. For those classes that do not explicitely need a
context manager to be used, the `FeatureContainer` context manager is
entered during initialization. It is never exited, as the classes do not
provide appropriate close methods or similar. This means the feature
accessors will only be removed during deletion of the instance itself
which should be fine.