Calculate image size by known values
The old calculation based on the pointer offsets could be problematic in
sitations where the assumption that chunk data is always placed before
iamge data in the buffer is incorrect. In situations where the chunk
data is added to the end of the buffer, the image size would include
this data and lead to incorrect interpretation of the memory.