Add context handling for TL feature access methods
The features should only be accessible, while the `VmbSystem` context is
open. When the `VmbSystem` context is closed, the
`TransportLayer._close` method is called. This leaves the context for
the `TransportLayer` instance and removes the detected and attached
features. The decoration added in this commit now ensures, that trying
to access the features via the inherited methods raises a
`RuntimeError`. The new implementation of the `RaiseIfOutsideContext`
decorator allows defining a custom message. As the relevant context for
the `TransportLayer` class is managed by the `VmbSystem` context, the
raised error message references that.