Add context sensitivity to `Stream`
The `Stream` context is handled via the `open` and `close` methods.
These are not marked as private because for devices with multiple
streams the user is supposed to be able to handle these manually. At the
same time, the stream at index 0 is the "default" stream that is used
when the streaming related `Camera` features are used. `Camera` will
handle opening and closing the stream at index 0 automatically.

Because of this the error message references the Cameras context. For
later implementations it might make sense to rethink this error message,
specifically when situations arise where users are allowed to close the
`Camera` context, but keep streams at index >=1 open.