Fix mypy error for `Camera.__local_device`
The typing for this variable is slightly tricky. The variable contains a
single instance of `LocalDevice`, but this instance only becomes
available when the context for `Camera` is entered (`_open` is called).
This means that during the call of `__init__`, the variable cannot be
assigned properly. To still get general type hints correct, the assigned
`None` in this line is excluded from type checking.