Use `filter_features_by_name` to get `StreamBufferAlignment`
The old implementation of simply trying to get the feature by name also
worked. But it had the side-effect, that the failed lookup of that
feature was added as an error to the log output. Using
`filter_features_by_name` does not add any error messages to the log
output if the feature is not found.