Fix gigantic timeout value in test case
The large timeout value was added for manual testing but does not make
sense as part of the test suite. 5 Seconds should be more than enough
for this test case.

The camera ID returned by `EventCameraDiscoveryCameraID` should be the
extended ID. This was not checked correctly and is also fixed in this
commit.