Turn context decorators back into classes
This is largely a revert of commit
d95995ea77dc79d6ecc5491629f79a144172b45a1

The reason for reverting this change is the newly introduced optional
`msg` parameter for the `RaiseIfInsideContext` and
`RaiseIfOutsideContext` decorators. Implementing the decorators as a
class makes this optional parameter easy to handle. The
`EnterContextOnCall` and `LeaveContextOnCall` decorators were changed
back to classes to keep a consistent style in the implementation.