Define callback function signatures in `vmb_c.py`
This better encapsulates required knowledge about the VmbC API to the
file `vmb_c.py`. Before, the type of function arguments for the
different callback functions was scattered in `feature.py` and
`stream.py`. Now these files simply include a pre-set constructor-like
type, that can be used to create the callbacks that are passed to VmbC.

Reordering of the file took place, because in order to work os-specific,
the VmbC library instance must already be loaded. After the instance is
loaded, the function signatures can be determined (depending on the type
of API, the funcptr type is slightly different) and the function
signatures attached.

This commit also already introduces the callback type and VmbC function
signature for accessing chunk data.