Reset user chunk exception so it cannot confuse us
Clearing the variable prevents a situation where an exception is raised
on some chunk access, handled as expected and some time later that
variable is looked at again and an exception found. The code could not
be sure, that this exception was already handled. By removing it we are
preventing this (very theoretical) situation.