Fix `Frame.convert_pixel_format`
This methods still tried to copy over ancillary data from the old frame
over to the new one. This is no longer the desired behaviour as
accessing that chunk data would be impossible anyway. In order to access
the chunk data, the `VmbFrame` instance would habe to be announced to
VmbC before calling `Frame.access_chunk_data`. This does not make sense
and therefore the chunk data can be ignored in the newly created frame.