Remove outdated TODO comment
My assumption that chunk callbacks are only allowed from within frame
callbacks was incorrect. This limitation currently exists in VmbPy (this
is why it is also documented as part of the methods docstring), but only
due to limitations in the synchronous acquisition procedures.

VmbPy records frames synchronously by entering capturing mode, waiting
for a frame to arrive, creating a copy of the frame, leaving capturing
mode and yielding the frame copy. This frame copy was never announced to
VmbC so trying to access chunk data associated with it will not work. To
enable chunk access with synchronous frame acquisition a major rework
and testing of the synchronous acquisition procedure is necessary.