Do not expose log level parameter
The issue with a log parameter would be, that we do not yet expose any
log level setting for the C-API. There is a log configuration procedure,
but that is only concerned with VmbPy logging and logs "end" at the VmbC
boundary. Introducing a log level parameter here might lead to confusion
regarding multiple different log types and the fact that the log entries
will not be added to the VmbPy log.

We will (for now) not expose the log level parameter via VmbPy. It was
not available in VimbaPython so this is not a step backwards. Might be
added in the future.

Discussed in Developer Alignment on 2022-08-12