Create tmpdir for load/save tests
If the test case is executed from a directory that does not already
contain another directory with the name 'Test', the changed test cases
failed due to path issues. By creating a temporary directory to store
the file in we can make sure that the test executes as expected without
having to worry about the working directory from which it is started.