Not relevant as VimbaX is only released as 64bit
The calculation itself does work as expected when using a 32bit Python
installation. However loading of VmbC.dll does not work, as the 32bit
Python process is not able to load a 64bit dll. If we ever decide to
compile VmbC as 32bit, the global VmbC Handle that is calculated here
would be calculates identically in VmbC so loading the new 32bit library
should be possible with a 32bit Python.
I expect that it would be possible to load a 32bit VmbC with a 64bit
Python, but the global VmbC handle would then not match meaning we would
not be able to use the actual functionality of VmbC. This is however
untested as there are currently no plans to compile VmbC as 32bit
library.