Iteration of `streamHandles` requires "casting"
While it is possible to iterate over ctype arrays is possible, the
`streamHandles` member is a void pointer. It would need to be "casted"
to a proper ctypes array for which the length of the array
(`streamCount`) needs to be attached to the type of elements in the
array. Doing all this just to be able to use a more pythonic loop here
does not seem worth the effort. The existing loop is readable enough as
is.