Fix attempted `get_id` call in `Stream.open`
This logged message was migrated from the `Camera` class. Since `Stream`
does not have a `get_id` method, it seems to make sense to log the
camera ID. Alternatively it would be possible to query some Stream
feature to get an ID (e.g. `StreamID` which is available with the
AVTGigETL), but I could not find a reference to a standardized feature
to rely on in the SFNC. Using the Camera ID seems safer in this case.