Add E402 to ignored flake8 errors for Tests
The E402 complains about module level imports that are not at the top of
the file. In our tests, this is done in order to insert the relative
path of VmbPy into the `sys.path` before importing `VmbPyTestCase`. This
is necessary because the `VmbPyTestCase` class itself is not really part
of VmbPy, but only the test helpers module.