Catch errors during tl/inter/cam initialization
If e.g. the transport layer does not behave as expected, some VmbC
function calls can fail during initialization of the instance. This was
observed for calls to `VmbFeaturesList` which are done when the feature
accessors are attached to the instances.

This logs the occuring errors to the VmbPy logger. The user must enable
logging in order to see the log messages. Slowdown of failing function
calls can still be critical.