Only encode file_path if it contains a value
The default case for VmbPy is that the file_path is not set to anything.
This means `None` is passed to `as_vmb_file_path`. `None` does obviously
not provide an `encode` method. With this change if `None` or an empty
string are passed to this function, `None` is returned from it. Ctypes
then handles this as a nullptr. This is the expected behaviour.