Fix string feature set test
The `get_max_length()` method returns the length that is actually
writeable for the feature. If we want to test the expectation 2
described in the test case, we need to make sure, that we try to write
*more* than the returned value.

This mistake was probably introduced, when the discussion about max
string length behaviour was last held. There the decision was made, that
the max length should be the length that is actually writeable. Before
in VimbaPython the last character was reserved for a \0 byte. So trying
to write the max length in the old implementation did not leave any room
for that, resulting in the expected error.