Rename `run_tests.py` to `Tests/__main__.py`
This allows using the entier `Tests` directory directly as essentially a
single python script. This in turn allows us to simply package the
content of `Tests` as a zip file, which can then be executed by Python
directly. This should make it very easy to provide the test suite to PQ
for validation.