Revert most changes
This did not seem to work as intended. It would make more sense to make
`Tests/__main__.py` a thin wrapper around `Tests/runner.py`. This way
`run_tests.py` would remain as is to be used as part of a future CI
pipeline and the Tests directory packaged as a zip file could be used by
PQ for validation.

This reverts the following commit
- b559dded278c8bc7ef417b3ffe5a4d2bec8b12e9
- abad5058d2cd4f6f2cdd2853665dd3e6fbb1eb1f
- 64c19cb4390affb3a7305f15df55d62fdd9cb1b2
- 4a0b36a5ba9b7478270c4489fa09f068fbd97314