Reset path configuration after test case
The changed test case sets the path configuration to an invalid number
to deliberately produce an error upon entering the `VmbSystem` context.
This invalid value for the path configuration must be manually reset to
the default value (`None`) to make sure that subsequent tests are not
blocked by the invalid value. If it is not reset all subsequent tests
will encounter an error when they attempt to enter the `VmbSystem`
context.