Use partial rollback for frame generator
This now only rolls the state machine back to the point where
`VmbCaptureEnd` has been called (in our experiments with VmbC before
this that seemed to be the minimum state we need to move back to to get
reliable synchronous acquisition done with USB and GigE cameras).

This also replaces the `frames` tuple that only contained a single frame
with a simple `frame` variable. This is wrapped into the length 1 tuple
again to be passed into the `_CaptureFsm` `_Context` argument as that
requires the frames to be iterable. Having a single `frame` variable
seems more convenient in the generator...