Make sure `get_frame` returns frame copy
`get_frame` does not make it possible to grant access to the original
frame to the user, as we have no way of knowing when the user is done
using the frame. Since the API might free the frame memory when a frame
is revoked (e.g. with allocation mode `AllocAndAnnounce`) we need to
create a copy that is handed to the user.

Side effect: The user will not be able to use a frame received with
`get_frame` to access chunk data for that frame. If chunk access is
required the user must use the new `get_frame_with_context` method.