Add type hints for mypy
The exclusion of the `forward` and `backward` call are necessary,
because the input `new_state` may be of a type that does not necessarily
implement `forward` or `backward` (this case is handled by the `except
AttributeError`). The complaint by mypy is valid, but we know that this
behaviour is intended so the complaint is explicitely ignored.