Improve exception handling in CaptureFsm
The state transitions now no longer return the next state. Instead each
state can be seen as more or less independent. A state can be `enter`ed,
which executes an appropriate VmbC function, and can be `exit`ed, which
returns the appropriate VmbC counterpart.

The entered states are stored as an array in the `CaptureFsm`. The
"current" state of the state machine corresponds to the last state that
was entered, i.e. the last element in `self.__states`.

The `enter` and `exit` methods of the states now also no longer catch
and return exception types. Instead exceptions are raised as usual. The
implementation of `go_to_state` however ensures, that all state
transitions that are expected will be attempted even if an exception is
encountered. See the docstring of that method for an explanation of how
exceptions are handled.

Changes to synchronous frame generator:
If an exception is encountered during the transition to `StateQueued`
after the frame has been received, that exception will be raised *after*
the frame was yielded to the user. This means that for context managed
access to the frame, the user will get the frame instance, and upon
leaving the context will be presented with the exception. If
non-context-managed method is used, the user will not receive the frame
instance but instead be presented only with the exception.