Fix `leave_capturing_mode`
The old implementation rolled back to the first state that is being
entered. This was correct while we still had the `Init` state, but is no
longer the desired behaviour. With the old implementation, the
`Announced` state is never exited, so frames are never revoked.

Instead the state machine should end up with an empty `__stated` array.
Making the `new_state` parameter an `Optional` allows it to clearly
communicate this possibility. If `None` is given as argument, the state
machine will exit all currently entered states as expected.