Fix chunk tests with real camera
For real cameras chunk features need to be individually enabled. This
enables the chunk features for each test case in the class, and disables
them after test execution. If a test case relies on chunk features being
disabled, they should be disabled explicitely in the test case or
implemented in a different test class.