Add failing test
This shows two different errors.

For allocation mode `AnnounceFrame`, a buffer is allocated and therefore
exists. With that mode a nullptr access will be reported when the frame
copy is attempted because `Frame._frame.imageData` is a nullptr.

For allocation mode `AllocAndAnnounceFrame`, no buffer is allocated by
VmbPy. This means that the copy already fails earlier when the
`Frame._buffer` variable is accessed. Here an `AttributeError` is raised
because the frame does not have a `_buffer` attribute.