Manually set `imageData` if it is not set by C-Level
This assumption might not be correct in all situations (in fact it is
probably only necessary to do this when no image data has been written
to the buffer), but it prevents nullptr evaluations and therefore
improves stability. The image that will be shown as result of this, is
likey all 0-bytes, or random noise depending on if the memory is nulled
during allocation or not.