Make `limit==0` a ValueError for `get_frame_generator`
If `limit=0` is accepted as allowed value, the frame generator starts
the camera stream by running though the entire capture FSM and closes it
again immediately.  For some camera models this seems to be problematic
(see below). By forbidding this value, we detect that no frames are
actually expected from this call (why the user would call with a limit
of 0 is not really clear anyway) and we therefore do not actually need
to start the camera acquisition.

With the old behaviour, it was possible to bring cameras into a state,
where streaming is blocked until a powercycle of the camera is
performed. The `AcquisitionStart` followed immediately by
`AcquisitionStop` seems to lead to a state, where acquisition is in some
kind of limbo state. The LED on the back of the device keeps blinking as
if a stream was running, but no frames are ever received. Repeatedly
calling `AcquisitionStop` does not have any impact on this. The device
must be power cycled (or DeviceReset must be executed) to get the camera
back to a working state. It is still possible to reach this state even
with this fix by manually calling `cam.start_streaming` and immediately
afterwards call `cam.stop_streaming`. I was not able to trigger the
state by running such a python script with an attached debugger, but
only by explicitely executing it with a python interpreter. It might be
possible, that this behaviour is *very* time sensitive and the debugger
or log statements add enough delay to the execution, that the camera
does not reach this state.

Setup that showed the error:
- Win 10
- Alvium 1800 U-1620c
- FW 11.0.9CF0C21E
- Failing test case:
  `real_cam_tetsts.camera_test.CamCameraTest.test_camera_frame_generator_limit_set`