Add small sleeps to dangerous test cases
These test cases do very little between stream start and stop. I
actually ancountered the problem described in the previous commit with
one of these tests as well. This delay is short enough to not blow up
the execution time too much, but at the same time hopefully prevents the
issue from coming up again. It is of course not ideal, but the quickest
fix available.