Use correct `_handle` attribute of `Interface`
The old `__handle` name that was used in VimbaPython has been changed to
`_handle` with PR #10. This change was not correctly implemented in the
interface callback function. When new interfaces were connected, this
lead to an `AttributeError` since the `Interface` instance did not have
a name mangled `__handle` (becomes `_Interface__handle`). This fixes
that error and allows interfaces that are added at runtime to be
correctly added to the list of detected interfaces by the `VmbSystem`.