Use `Camera._disconnected` to check feature access
This variable gets set to `True` by the camera change callback when the
camera connection goes missing. This usually happens if the camera gets
disconnected. If the camera is disconnected, accessing features of the
remote device or stream will not work for obvious reasons.

The variable is used in the capture state machine, to check if we expect
feature access to work before running the `AcquisitionStart` and
`AcquisitionStop` command features. On the setup phase, losing the
connection raises a Camera error to indicate to the user, that the
camera is not accessible. On the shutdown phase of the stream, if the
camera is not accessible the shutdown simply skips the `AcquisitionStop`
execution. This might not be ideal, but without a reliable camera
connection the feature execution would result in an exception and
probably not be performed on the device anyway.