Use `Camera._disconnected` in `is_streaming`
The `_disconnected` variable was used in VimbaPython to check if the
camera was streaming before. It was removed during the migration of the
streaming implementation to `stream.py`. Since the `_disconnected`
variable is now being used again in the capture state machine logic, it
also makes sense to use it again to check if a camera/stream is
streaming.