Add `Frame._buffer` to numpy dtype metadata
This adds a reference to the buffer used to store image data, to ensure
that even if the original VmbPy `Frame` instance is garbage collected,
the buffer does not get freed until the numpy array is also garbage
collected. Otherwise accessing the numpy array after the `Frame` was
garbage collected lead to a segfault because the freed memory would be
accessed.