Replace docopt with arpgarse for CLI arguments
This removes the third-party dependency docopt and instead uses the
standard library module argparse. Functionality is the same as before,
but the command line interface slightly changed. The reporting is now no
longer set via the -o flag, but instead by having mutually excusive
options (`--console` and `--junit_xml`).

The CLI for `run_tests.py` is now also harmonized with that of
`runner.py` and the more-or-less same CLI from before is no longer
duplicated in two places. Instead the CLI is simply imported from
runner.py.