Set pixel format to compatible value if possible
This check to see if the camera supports a numpy compatible pixel format
and if so makes sure a compatible format is set on the device before the
frame is acquired. This checks agains the `OPENCV_PIXEL_FORMATS` because
they are already conveniently available, but generally more formats can
be represented as numpy array. This should still be enough for most of
our devices and is easier to read than an alternative implementation
(see `as_numpy_ndarray` to check how validity of the format is checked
there).