Reset pixelformat after test case execution
The `convert_to_all_give_formats` test case sets all possible pixel
formats of the camera. This means that by the end of the test case, the
pixel format that was set at the beginning might be changed. This can
have an impact on streaming tests e.g. when a higher bit-depth format is
set than expected. Ideally other test cases would check the set pixel
format and if necessary set it to a desired value, but this is not
always done. Resetting the pixel format to the initial value is not
dangerous here and might reduce confusion down the road.