Add tox.ini experiment
This file can be used to execute the test suite via pytest and generate
a coverage report, as well as a junit xml file with the test results.
These files are placed in a directory, that is named after the
environment in which the test execution took place. This should make it
easier to identify, which environment had issues with the test
execution.

The benefit of using tox is that it automatically handles installation
of the module under test (in our case VmbPy) as well as the listed
dependencies into a new test-environment.

Another great benefit of tox is the option, to easily add more
environments that should be tested against. This can for example be used
to execute the test suite with multiple python versions, or multiple
versions of third party libraries.