Remove incorrect timeout calculation
The removed section of the code attempted to calculate the required time
to record images from the exposure time. This was wrong for two reasons:
1. ExposureTime is not reported in seconds. An appropriate conversion
   to seconds was missing meaning the calculated fps were not measured
   in hetz as expected
2. Even if a camera records frames with a short exposure time, it is not
   guaranteed that frames will be transmitted just as quickly.
   Especially for very short exposure times and cameras with large
   sensores the transmission time will become a limiting factor in the
   achievable frames per second.

If the camera does not provide the `AcquisitionFrameRate` feature, the
function will no longer try to determine a fallback value based on the
exposure time.

Co-authored by: Simon Albino <Simon.Albino@alliedvision.com>