Fix uses of `get_frame_generator` in test suite
`get_frame_generator` produces frames, that may only be used inside
their respective loop iteration. Most of these uses were not really
critical, but some of them "collected" frames into a list and stored
that beyond the actual lifetime guarantee of the frame. This was changed
so as not to provide incorrect usage examples for people who look at the
test suite.