Change `max_iterations` default for save to `0`
This makes use of the new behavior of VmbC where passing a `0` for
`maxIterations` to `VmbSettingsSave` will tell VmbC to auto select a
value (currently defaults to 5). That value is written to the XML file
and if in `VmbSettingsLoad` the `maxIterations` field is set to `0`, the
value from the XML is used.

Also adjusts the dosctrings slightly to point out this behavior.